updated i18next and Translation load order
