+++ client/modules/i18n/templates/header/i18n.js
@@ -1,7 +1,7 @@
 import { Reaction } from "/client/api";
 import { Shops } from "/lib/collections";
 import { Session } from "meteor/session";
-
+import { i18nextDep } from "../../main";
 /**
  * i18nChooser helpers
  */
@@ -23,11 +23,13 @@ Template.i18nChooser.helpers({
         }
       }
     }
+    return languages;
   },
   active() {
     if (Session.equals("language", this.i18n)) {
       return "active";
     }
+    return "";
   }
 });
 
@@ -38,6 +40,12 @@ Template.i18nChooser.helpers({
 Template.i18nChooser.events({
   "click .i18n-language"(event) {
     event.preventDefault();
+    //
+    // this is a sanctioned use of Meteor.user.update
+    //
+    // console.log("this.i18n", this.i18n)
+    Meteor.users.update(Meteor.userId(), {$set: {"profile.lang": this.i18n}});
+
     return Session.set("language", this.i18n);
   }
 });
